split io mod
